package_wip: Add kernel oci runtime-spec object
This commit adds a generated oci runtime-spec kernel object to the
archive filesystem at /rootfs/vm/vm.json. The vm.json file describes
path to a kernel image and parameters for booting the kernel along with
additional artifacts such as initrd and filesystems. This is needed
because containerd creates a container specific runtime config. During a
runtime create method, a runtime should use the vm object as a template
a bundles runtime-spec config object.