add fake command for containerd
Containerd expects to create a container, as such when generating a
runtime-spec config object, containerd requires a command from the
image-spec config to be used as init process inside a contianer. Runu
creates uk vm's and as such does not specify a command. We simply add a
'dummy' command satisfying containerd's requirements.